backend/requirements.txt
Cython>=0.29.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
opencv-python-headless==4.8.1.78
pysaliency==0.2.19
scipy==1.11.3
numpy==1.26.2
google-generativeai==0.3.1
fpdf2==2.7.5
python-dotenv==1.0.0
